Crud+ jwt

+++ controllers/contactContoller.js
@@ -2,16 +2,17 @@ const asyncHandler = require("express-async-handler");
 const  Contact = require("../models/contactModel")
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    //getting the contacts from the db
-    const contacts = await Contact.find();
+  //getting the contacts from the db
+  //getting all the contacts created by the login in user
+  const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
   });
 
 //@desc Post contact by id
 //@routs POST /api/contacts/
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     //printing the request body
     console.log("The request body is : ", req.body);
@@ -27,14 +28,16 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
       name, //equls to request.body.name. in es6 key and value is same we can use key
       email,
-      phone
+      phone,
+      user_id: req.user.id, //req.user will come from the middleware
+
     });
     res.status(201).json(contact);
   });
 
 //@desc Get new contacts
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     //getting the contact by id
     const contact = await Contact.findById(req.params.id);
@@ -47,14 +50,18 @@ const getContact = asyncHandler(async (req, res) => {
   });
 //@desc Update all contacts
 //@route UPDATE /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
       res.status(404);
       throw new Error("Contact not found");
     }
-  
+    //checking weather the user id of the login user is matching with the updating contact
+    if (contact.user_id.toString() !== req.user.id) {
+      res.status(402);
+      throw new Error("User don't have permission to update this");
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
       req.params.id, //getting the id of the contact that needs to be updated
       req.body, //getting the updated body
@@ -70,14 +77,19 @@ const updateContact = asyncHandler(async (req, res) => {
 //@access public
 //@desc Delete a contact
 //@routs DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
       res.status(404);
       throw new Error("Contact not found");
     }
-  
+    //checking weather the user id of the login user is matching with the deleting contact
+    if (contact.user_id.toString() !== req.user.id) {
+      res.status(402);
+      throw new Error("User don't have permission to update this");
+    }
+    
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact); 
   });
